requirements.txt
Django==4.2.6
djangorestframework
django-filter
psycopg2==2.9.9
python-dotenv
drf-spectacular